chore: add .gitattributes to enforce LF line endings chore: 改行コードをLFに統一する.gitattributesを追加

+++ src/db/models.rs
@@ -1,19 +1,17 @@
 // Database model structs
 
-use serde::Deserialize;
-
-#[derive(Clone, Deserialize)]
+#[derive(Clone, sqlx::FromRow)]
 pub struct AuthUser {
     pub id: i32,
     pub email: String,
     pub display_name: String,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct Quiz {
     pub id: i32,
     pub name: String,
-    pub count: i32,
+    pub count: i64,
 }
 
 pub struct QuestionModel {
@@ -22,7 +20,7 @@ pub struct QuestionModel {
     pub options: Vec<QuestionOptionModel>,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct QuestionOptionModel {
     pub id: i32,
     pub is_answer: bool,
@@ -30,7 +28,7 @@ pub struct QuestionOptionModel {
     pub explanation: Option<String>,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct QuizSessionModel {
     pub id: i32,
     pub quiz_id: i32,
@@ -39,7 +37,7 @@ pub struct QuizSessionModel {
     pub selection_mode: Option<String>,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct AnswerModel {
     pub question: String,
     pub is_correct: bool,
@@ -47,13 +45,13 @@ pub struct AnswerModel {
     pub is_bookmarked: bool,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct QuestionStatsModel {
     pub question: String,
-    pub correct_answers: i32,
+    pub correct_answers: i64,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct SessionReportModel {
     pub id: i32,
     pub name: String,
@@ -66,33 +64,33 @@ pub struct SessionReportModel {
     pub selection_mode: Option<String>,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct CategoryStats {
     pub category: String,
-    pub total: i32,
-    pub correct: i32,
+    pub total: i64,
+    pub correct: i64,
     pub accuracy: f64,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct QuizOverallStats {
-    pub total_questions: i32,
-    pub unique_asked: i32,
-    pub total_correct: i32,
-    pub total_answered: i32,
+    pub total_questions: i64,
+    pub unique_asked: i64,
+    pub total_correct: i64,
+    pub total_answered: i64,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct DailyAccuracy {
     pub date_label: String,
     pub accuracy: f64,
 }
 
-#[derive(Deserialize)]
+#[derive(sqlx::FromRow)]
 pub struct QuizCategoryOverallStats {
     pub category: String,
-    pub total_in_category: i32,
-    pub unique_asked: i32,
-    pub total_correct: i32,
-    pub total_answered: i32,
+    pub total_in_category: i64,
+    pub unique_asked: i64,
+    pub total_correct: i64,
+    pub total_answered: i64,
 }
